+++ changelog.md
@@ -1,3 +1,9 @@
-# 0.1.0 - spa embedding into go
+# 0.1.1 - Update: Dockerfile and fake loading
+- fake loading removed in frontend
+- dockerfile now updated to work on fly.io
+- fly.io deployment working
+- updated readme and claude.md
+
+# 0.1.0 - Add: spa embedding into go
 - mock frontend created
 - initial commit for the spa served in go web app